Refactor: rename all component
