+++ frontend/index.html
@@ -5,7 +5,7 @@
 		<link
 			rel="icon"
 			type="image/svg+xml"
-			href="/vite.svg"
+			href="https://static.chotot.com/storage/careers/2021/10/chotot-square.png"
 		/>
 		<meta
 			name="viewport"
