+++ frontend/src/components/commons/header/index.ts
@@ -1,4 +1,4 @@
-export * from './header.component'
-export * from './header-top.component'
-export * from './header-bottom.component'
-export * from './header-search.component'
+export * from './Header'
+export * from './HeaderTop'
+export * from './HeaderBottom'
+export * from './HeaderSearch'
