+++ frontend/src/components/commons/navbar/Navbar.tsx
@@ -1,0 +1,6 @@
+export function Navbar() {
+	return (
+		<nav className='bg-app-primary'>
+		</nav>
+	)
+}
