+++ frontend/src/components/commons/navbar/index.ts
@@ -1,1 +1,1 @@
-export * from './navbar.component'
+export * from './Navbar'
