--- frontend/src/components/commons/navbar/navbar.component.tsx
@@ -1,3 +1,0 @@
-export function Navbar() {
-	return <div className='bg-app-primary'></div>
-}
