+++ frontend/src/main.tsx
@@ -1,20 +1,20 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import LoginPage from './pages/login.tsx'
+import LoginPage from '@/pages/login.tsx'
+import RegisterPage from '@/pages/register.tsx'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-import RegisterPage from './pages/register.tsx'
 // import { AppProvider } from '@/components/context/app.context.tsx'
 // import ProtectedRoute from '@/components/auth/index.tsx'
 import { Toaster } from './components/ui/sonner.tsx'
 // import { SocketProvider } from './components/context/socket.context.tsx'
-import HomePage from './pages/home.tsx'
+import HomePage from '@/pages/home.tsx'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import Layout from './layout/layout.tsx'
+import Layout from '@/layout/layout.tsx'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
-import UpdateProfile from './pages/update.profile.tsx'
-import AccountSettings from './pages/account.setting.tsx'
-import { ProfileLayout } from './layout/layout.profile.tsx'
+import UpdateProfile from 'pages/UpdateProfile.tsx'
+import AccountSettings from 'pages/AccountSetting.tsx'
+import { ProfileLayout } from '@/layout/LayoutProfile.tsx'
 
 const router = createBrowserRouter([
 	{
