+++ frontend/src/services/api/message.ts
@@ -1,0 +1,5 @@
+import axios from "@/services/AxiosCustomize";
+
+export const getMessagesAPI = (receiverId: string) => {
+  return axios.get<IBackendRes<IMessage[]>>(`/api/messages/${receiverId}`);
+};
