+++ frontend/src/services/sleep.ts
@@ -1,0 +1,1 @@
+export const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
