+++ frontend/src/types/global.d.ts
@@ -1,73 +1,5 @@
-export { };
-
-declare global {
-    interface IBackendRes<T> {
-        error?: string | string[];
-        message: string;
-        statusCode: number | string;
-        data?: T;
-    }
-
-    interface IModelPaginate<T> {
-        meta: {
-            current: number;
-            pageSize: number;
-            pages: number;
-            total: number;
-        };
-        result: T[];
-    }
-
-    interface ILogin {
-        access_token: string;
-        user: {
-            _id: string;
-            name: string;
-            email: string;
-        };
-    }
-
-    interface IRegister {
-        _id: string;
-        createdAt: string;
-    }
-
-    interface IUser {
-        email: string;
-        name: string;
-        _id?: string;
-    }
-
-
-    interface IUserUpdatePayload {
-        id: number;
-        name: string;
-        email: string;
-    }
-    interface IFetchAccount {
-        user: IUser;
-    }
-
-    interface IUserTable {
-        _id: string;
-        name: string;
-        email: string;
-        createdAt: string;
-        updatedAt: string;
-    }
-
-    export interface IMessage {
-        _id: string;
-        senderId: string;
-        receiverId: string;
-        message: string;
-        createdAt: string;
-        updatedAt: string;
-    }
-
-    export interface IOnlineUser {
-        online: string[]
-    }
-
-
-}
+/// <reference path="./user.d.ts" />
+/// <reference path="./api.d.ts" />
+/// <reference path="./message.d.ts" />
+/// <reference path="./files.d.ts" />
+/// <reference path="./axios.d.ts" />
