+++ frontend/src/types/message.d.ts
@@ -1,0 +1,12 @@
+ interface IMessage {
+  _id: string;
+  senderId: string;
+  receiverId: string;
+  message: string;
+  createdAt: string;
+  updatedAt: string;
+}
+
+ interface IOnlineUser {
+  online: string[];
+}
